# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.0
paths:
...
  /api/v1/faction-relations:
    get:
      tags:
      - factions
      summary: List All Faction Relations
      description: Every faction-to-faction relation from the active save. Returns
        [] until a save is ingested.
      operationId: list_all_faction_relations_api_v1_faction_relations_get
      parameters:
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          description: Include utility/hidden factions.
          default: false
          title: Include Hidden
        description: Include utility/hidden factions.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AllFactionRelation'
                title: Response List All Faction Relations Api V1 Faction Relations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/factions:
    get:
      tags:
      - factions
      summary: List Factions
      description: 'List factions in the game catalog. Overrides the player faction
        name with the

        custom organisation name from the save when a save has been ingested.'
      operationId: list_factions_api_v1_factions_get
      parameters:
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          description: Include utility/hidden factions.
          default: false
          title: Include Hidden
        description: Include utility/hidden factions.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactionSummary'
                title: Response List Factions Api V1 Factions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/factions/strength:
    get:
      tags:
      - factions
      summary: Faction Strength
      description: 'Relative strength metrics, normalized 0-100, computed from LIVE
        save state.


        Everything here is the *current* universe (dynamic ships/stations/relations)
        so the

        standings reflect how the game has actually unfolded — seed/static is init
        state only

        and is not referenced once a save is loaded. The player is ranked alongside
        AI factions.'
      operationId: faction_strength_api_v1_factions_strength_get
      parameters:
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          description: Include utility/hidden factions.
          default: false
          title: Include Hidden
        description: Include utility/hidden factions.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactionStrength'
                title: Response Faction Strength Api V1 Factions Strength Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/factions/known:
    get:
      tags:
      - factions
      summary: List Known Factions
      description: 'Return {faction_id: is_known} for every static faction.'
      operationId: list_known_factions_api_v1_factions_known_get
      parameters:
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          description: Include utility/hidden factions.
          default: false
          title: Include Hidden
        description: Include utility/hidden factions.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
                title: Response List Known Factions Api V1 Factions Known Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/factions/{faction_id}:
    get:
      tags:
      - factions
      summary: Get Faction
      description: Get detailed information for a specific faction.
      operationId: get_faction_api_v1_factions__faction_id__get
      parameters:
      - name: faction_id
        in: path
        required: true
        schema:
          type: string
          title: Faction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FactionDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/factions/{faction_id}/relations:
    get:
      tags:
      - factions
      summary: List Faction Relations
      description: Diplomatic relations for a faction from the active save. Returns
        [] until a save is ingested.
      operationId: list_faction_relations_api_v1_factions__faction_id__relations_get
      parameters:
      - name: faction_id
        in: path
        required: true
        schema:
          type: string
          title: Faction Id
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          description: Include utility/hidden factions.
          default: false
          title: Include Hidden
        description: Include utility/hidden factions.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactionRelation'
                title: Response List Faction Relations Api V1 Factions  Faction Id  Relations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/licences:
    get:
      tags:
      - factions
      summary: List Licences
      description: List all faction licences, optionally filtered by faction_id.
      operationId: list_licences_api_v1_licences_get
      parameters:
      - name: faction_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactionLicence'
                title: Response List Licences Api V1 Licences Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AllFactionRelation:
      properties:
        faction_id:
          type: string
          title: Faction Id
        other_faction_id:
          type: string
          title: Other Faction Id
        initial_relation:
          type: number
          title: Initial Relation
        current_relation:
          anyOf:
          - type: number
          - type: 'null'
          title: Current Relation
      additionalProperties: false
      type: object
      required:
      - faction_id
      - other_faction_id
      - initial_relation
      title: AllFactionRelation
...
    FactionDetail:
      properties:
        faction_id:
          type: string
          title: Faction Id
        name:
          type: string
          title: Name
        color_hex:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Hex
        is_hidden:
          type: boolean
          title: Is Hidden
          default: false
        short_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Short Name
        prefix_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Prefix Name
        space_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Space Name
        home_space_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Home Space Name
        police_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Police Faction
        primary_race:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Race
        icon_active:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Active
        icon_inactive:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Inactive
        icon_banner:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Banner
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        behaviour_set:
          anyOf:
          - type: string
          - type: 'null'
          title: Behaviour Set
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
      additionalProperties: false
      type: object
      required:
      - faction_id
      - name
      - color_hex
      - primary_race
      title: FactionDetail
    FactionLicence:
      properties:
        licence_type:
          type: string
          title: Licence Type
        faction_id:
          type: string
          title: Faction Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        precursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Precursor
        price:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price
        min_relation:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Relation
      additionalProperties: false
      type: object
      required:
      - licence_type
      - faction_id
      title: FactionLicence
    FactionRelation:
      properties:
        other_faction_id:
          type: string
          title: Other Faction Id
        initial_relation:
          type: number
          title: Initial Relation
        current_relation:
          anyOf:
          - type: number
          - type: 'null'
          title: Current Relation
      additionalProperties: false
      type: object
      required:
      - other_faction_id
      - initial_relation
      title: FactionRelation
    FactionStrength:
      properties:
        faction_id:
          type: string
          title: Faction Id
        name:
          type: string
          title: Name
        color_hex:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Hex
        military_score:
          type: number
          title: Military Score
        economic_score:
          type: number
          title: Economic Score
        diplomatic_score:
          type: number
          title: Diplomatic Score
        territory_score:
          type: number
          title: Territory Score
        military:
          $ref: '#/components/schemas/FactionStrengthBreakdown'
        economic:
          $ref: '#/components/schemas/FactionStrengthBreakdown'
        territory:
          $ref: '#/components/schemas/FactionStrengthBreakdown'
        diplomacy:
          $ref: '#/components/schemas/FactionStrengthBreakdown'
        fight_ship_count:
          type: integer
          title: Fight Ship Count
        trade_ship_count:
          type: integer
          title: Trade Ship Count
        mine_ship_count:
          type: integer
          title: Mine Ship Count
        military_station_count:
          type: integer
          title: Military Station Count
        economic_station_count:
          type: integer
          title: Economic Station Count
        sector_count:
          type: integer
          title: Sector Count
        cluster_count:
          type: integer
          title: Cluster Count
        avg_relation:
          type: number
          title: Avg Relation
      additionalProperties: false
      type: object
      required:
      - faction_id
      - name
      - color_hex
      - military_score
      - economic_score
      - diplomatic_score
      - territory_score
      - military
      - economic
      - territory
      - diplomacy
      - fight_ship_count
      - trade_ship_count
      - mine_ship_count
      - military_station_count
      - economic_station_count
      - sector_count
      - cluster_count
      - avg_relation
      title: FactionStrength
    FactionStrengthBreakdown:
      properties:
        raw:
          type: number
          title: Raw
        leader_raw:
          type: number
          title: Leader Raw
        leader_ratio:
          type: number
          title: Leader Ratio
        components:
          items:
            $ref: '#/components/schemas/FactionStrengthComponent'
          type: array
          title: Components
      additionalProperties: false
      type: object
      required:
      - raw
      - leader_raw
      - leader_ratio
      - components
      title: FactionStrengthBreakdown
    FactionStrengthComponent:
      properties:
        label:
          type: string
          title: Label
        value:
          type: number
          title: Value
        detail:
          type: string
          title: Detail
      additionalProperties: false
      type: object
      required:
      - label
      - value
      - detail
      title: FactionStrengthComponent
    FactionSummary:
      properties:
        faction_id:
          type: string
          title: Faction Id
        name:
          type: string
          title: Name
        color_hex:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Hex
        is_hidden:
          type: boolean
          title: Is Hidden
          default: false
        short_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Short Name
        prefix_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Prefix Name
        space_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Space Name
        home_space_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Home Space Name
        police_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Police Faction
        primary_race:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Race
        icon_active:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Active
        icon_inactive:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Inactive
        icon_banner:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Banner
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
      additionalProperties: false
      type: object
      required:
      - faction_id
      - name
      - color_hex
      title: FactionSummary
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...